Added service to hold the computation at background

+++ app/src/main/java/com/exaltead/sceneclassifier/data_extraction/IAudioBufferer.kt
@@ -2,5 +2,5 @@ package com.exaltead.sceneclassifier.data_extraction
 
 interface IAudioBufferer {
 
-    fun takeShortAudioRecord(duration: Float): ShortArray
+    fun takeShortAudioRecord(duration: Double): ShortArray
 }
